pyproject.toml
[tool.poetry]
name = "perseus"
version = "0.1.0"
description = ""
authors = ["Michael Roberts <michael@observerly.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.73.0"
uvicorn = {extras = ["standard"], version = "^0.17.4"}
pydantic = "^1.9.0"
email-validator = "^1.1.3"
sqlalchemy = "^1.4.31"
alembic = "^1.7.6"
astropy = "^5.0.1"
psycopg2 = "^2.9.3"
orjson = "^3.6.7"
furl = "^2.1.3"
requests = "^2.27.1"
sentry-sdk = "^1.5.10"
fastapi-cache2 = {extras = ["redis"], version = "^0.1.8"}
aioredis = "^2.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
flake8 = "^4.0.1"
autoflake = "^1.4"
isort = "^5.10.1"
black = "^22.1.0"
mypy = "^0.931"
tenacity = "^8.0.1"
httpx = "^0.22.0"
anyio = {extras = ["trio"], version = "^3.6.1"}
pytest-asyncio = "^0.18.3"
asgi-lifespan = "^1.0.1"

[tool.black]
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | venv
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black" 
multi_line_output = 3
include_trailing_comma = false
force_grid_wrap = 0
line_length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
